Move utils to utils file
